Create README.md
Semantic processing is the most challenging area in the field of NLP because it is difficult to let machines understand the text the same way as human do such as inferring the intent, distinguishing ambiguity of words, dealing with synonyms, detecting sarcasm, etc. Techniques for Semantic processing include the following: 1. Applications of knowledge graphs; 2. WordNet; 3. Word sense disambiguation; 4. Distributional Semantics; and 5. Topic Modelling.